+++ src/components/ui/index.ts
@@ -1,6 +1,6 @@
 import Button from "./button/button";
-import CategoryInput from "./input/CategoryInput";
-import GenderInput from "./input/GenderInput";
+import CategoryInput from "./input/category-input";
+import GenderInput from "./input/gender-input";
 import Input from "./input/Input";
 import Modal from "./modal/gathering-modal";
