+++ src/components/ui/input/Input.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useFormContext } from "react-hook-form";
-import StatusMessage from "./StatusMessage";
+import StatusMessage from "./status-message";
 
 interface InputProps {
   name: string;
