+++ src/components/ui/input/category-input.tsx
@@ -1,5 +1,5 @@
 import { CATEGORIES } from "@/constants/constants";
-import OptionInput from "./OptionInput";
+import OptionInput from "./option-input";
 
 interface CategoryInputProps {
   label?: string | React.ReactNode;
